backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.2
django-simple-search==1.0.2
djangorestframework==3.16.1
httplib2==0.31.0
pillow==12.0.0
pyparsing==3.2.5
python-dotenv==1.2.1
shortuuid==1.0.13
sqlparse==0.5.3
